+++ packages/mdx-test/content/with-links.mdx
@@ -18,3 +18,9 @@ Visit [Daring Fireball][] for more information. Or [Google][] it.
 [map]: https://clojuredocs.org/clojure.core/map "Map Docs"
 [range]: https://clojuredocs.org/clojure.core/range "Range Docs"
 [cycle]: https://clojuredocs.org/clojure.core/cycle "Cycle Docs"
+
+---
+
+TODO
+
+- fix underline color (maybe add .sass and use token's color)
