+++ packages/mdx/src/components.tsx
@@ -2,45 +2,12 @@ import React from "react"
 import { CodeSpring } from "@code-hike/smooth-code"
 import { EditorSpring } from "@code-hike/mini-editor"
 import { Code } from "@code-hike/utils"
+import { Section, SectionLink } from "./section"
 
 export const CH = {
   Code,
   Section,
   SectionLink,
-}
-
-const SectionContext = React.createContext({})
-
-function Section({
-  children,
-}: {
-  children: React.ReactNode
-}) {
-  return (
-    <section style={{ outline: "1px solid red" }}>
-      <div>{children}</div>
-    </section>
-  )
-}
-
-function SectionLink({
-  focus,
-  children,
-}: {
-  focus: string
-  children: React.ReactNode
-}) {
-  return (
-    <span
-      style={{
-        textDecoration: "underline",
-        textDecorationStyle: "dotted",
-      }}
-      title={focus}
-    >
-      {children}
-    </span>
-  )
 }
 
 function Code({
